Add RuboCop and correct/ignore offenses
This change imposes a style guide, that can be altered by updates to
`.rubocop.yml`, for use on this gem. Hopefully soon I'll be able to
access CodeClimate and set up automatic code style checks using the
provided RuboCop configuration. The purpose of adding a style guide to
this, and eventually all, redis-store repo(es) is to emphasize
readability, understandability, and to keep us up-to-date with the
latest "dialect" of Ruby that people are using. This helps new
developers as well as veterans get involved in our various projects.

Until we have some time to write documentation and solve some of the
other ignored RuboCop offenses, a few different style and lint checks
have been disabled so that we can continue working on the project and
not get bogged down by style errors that are not practical to fix at
this time.