Set client option `:ssl` to true when URI scheme is `rediss://`
When Redis::Store is initialized with a URI that has a scheme of
`rediss://`, the client option `:ssl` should be set to `true`.
However, the scheme option is deleted because the redis-client does not
have a `scheme` in the `_remove_unsupported_options` method.

Then Redis::Client only knows host, port, and password, and didn't
connect with SSL.
So if the passed URL has `rediss://` scheme, to tell Redis::Client
needs to connect with ssl make `:ssl` option to `ture`